requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
pydantic
python-dotenv
sentence-transformers
faiss-cpu
groq
passlib[bcrypt]
python-multipart
PyPDF2
pgvector